requirements.txt
psycopg2 == 2.9.10
django == 5.2
djangorestframework == 3.16.0

